move partitionCost to NetworkIngressCost
